+++ packages/rpc_dart_blob/lib/src/models.dart
@@ -1,5 +1,3 @@
-import 'dart:convert';
-
 import 'package:equatable/equatable.dart';
 import 'package:meta/meta.dart';
 import 'package:rpc_dart/rpc_dart.dart';