+++ packages/rpc_dart_data/test/data_service_rpc_test.dart
@@ -221,16 +221,4 @@ class _ThrowingRepository implements IDataRepository {
     // TODO: implement setSchemaPolicy
     throw UnimplementedError();
   }
-
-  @override
-  Future<StartMigrationResponse> startMigration(StartMigrationRequest request) {
-    // Not used in these tests.
-    throw UnimplementedError();
-  }
-
-  @override
-  Future<MigrationStatusResponse> getMigrationStatus(String collection) {
-    // Not used in these tests.
-    throw UnimplementedError();
-  }
 }